backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
pgvector
torch
torchvision
open-clip-torch
imagehash
Pillow
sentence-transformers
resend
jinja2
fpdf2
python-dotenv
httpx
playwright
google-genai
google-cloud-vision
google-cloud-aiplatform
google-auth
google-auth-oauthlib
pydantic-settings
